src/plugins/aidev_wxbot/pyproject.toml
[project]
name = "aidev-wxbot"
version = "1.0.0"
description = "A Django app for WeChat AI bot integration with BK AI Dev platform"
authors = [
    {name = "tencentblueking"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "django>=4.0,<5.0",
    "djangorestframework>=3.14.0",
    "pydantic>=2.11.7",
    "stackprinter>=0.2.12",
    "pycryptodome>=3.23.0",
    "requests>=2.32.5",
    "pika>=1.3.0,<2.0"
]

[project.scripts]
"aidev_wxbot.manage" = "aidev_wxbot.manage:main"


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["aidev_wxbot"]

[tool.ruff]
exclude = ["*/config/", "*/migrations/"]
force-exclude = true
line-length = 120
indent-width = 4

[tool.ruff.lint]
select = ["E", "F", "I", "PIE", "PERF", "SIM"]
ignore = ["PERF401", "E501"]

[tool.ruff.format]
quote-style = "double"

[tool.ruff.lint.isort]
# 总是显式制定 import section 的顺序
section-order = [
  "future",
  "standard-library",
  "third-party",
  "first-party",
  "local-folder",
]
relative-imports-order = "closest-to-furthest"

[tool.pytest.ini_options]
env_override_existing_values = 1
env_files = [".env"]
asyncio_default_fixture_loop_scope = "session"
addopts = "-vvv --maxfail=2 -s"
asyncio_mode = "auto"